requirements.txt
Appium-Python-Client
iniconfig==2.0.0
pip
pycparser==2.22
pytest==8.2.2
selenium==4.21.0
browserstack-local==1.2.10
jsonmerge
multiprocess
paver
psutil
pytest-variables
pytest-selenium
browserstack_sdk==1.23.6
pyautogui
Pillow
pytest-html
pytz
requests_ntlm
webdriver_manager